refactor(UI): :recycle: refactor UI methods
Refactored UI methods into a separate file for easier maintainability.
It is responsible for UI components such as the player informations and
the win/lose screen.
